Move Var to DAGCircuit.
